Distance font supports signed distance function in glsl

+++ src/WRATH/text/shaders/distance/font_distance_base.frag.glsl
@@ -1,6 +1,7 @@
 #define WRATH_FONT_IMPLEMENT_SIGNED_DISTANCE
 
-float wrath_font_distance_compute_signed_distance(float rr)
+mediump float 
+wrath_font_distance_compute_signed_distance(float rr)
 {
   return wrath_font_page_data(3)*(2.0*rr - 1.0);
 }
